refactor(404): refactor 404 to component for easier logic on dynamic slug page
